Created DBContext and Factory for design-time creation, added migration with some sample data

+++ TestBarberPWA/Shared/ServiceSold.cs
@@ -11,6 +11,6 @@ namespace TestBarberPWA.Shared
         public int AppointmentID { get; set; }
         public int ServiceID { get; set; }
         public int Quantity { get; set; }
-        public int SubTotal { get; set; }
+        public float SubTotal { get; set; }
     }
 }
